Remove btn_example.py
I'd like to remove that example, because it shows how to not handle
buttons. The problem is that there is the `gamepad` module running
in the background and checking the buttons for us, together with
debouncing and all that, and touching the button pins directly like
in this example may interfere with that in unpredictable ways.

I'm thinking about a way of hiding some of the "internal" pins, or
at least prefixing them with an underscore as to suggest that they
shouldn't be normally touched by the users without knowing what they
are doing.